third_party/blink/web_tests/platform/generic/fast/deprecated-flexbox/vertical-box-form-controls-expected.txt

Check if form controls in vertical flex box will stretch horizontally when rendered as box as opposed to inline-box. If you see any red, then the test has failed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Button
PASS element.offsetHeight is 100
PASS element.offsetWidth is 100
Text input
PASS element.offsetHeight is 100
PASS element.offsetWidth is 100
Textarea
PASS element.offsetHeight is 100
PASS element.offsetWidth is 100
Submit
PASS element.offsetHeight is 100
PASS element.offsetWidth is 100
PASS successfullyParsed is true

TEST COMPLETE
